// src/api/generated/models/teamRead.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SoftTrack
 * An open-source, self-hostable issue tracker inspired by Linear.
 * OpenAPI spec version: 0.1.0
 */

export interface TeamRead {
  id: number;
  name: string;
  key: string;
  description?: string | null;
  created_at: string;
}